@@ -3,22 +3,21 @@
   "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "flux": "^2.0.0",
-    "keymirror": "~0.1.0",
-    "object-assign": "^1.0.0",
-    "react": "^0.12.0",
-    "autolinker": "^0.15.2",
+    "autolinker": "^0.18.1",
+    "flux": "^2.1.1",
+    "keymirror": "^0.1.1",
+    "object-assign": "^3.0.0",
+    "react": "^0.13.3",
     "react-zeroclipboard-mixin": "^0.1.0"
   },
   "devDependencies": {
-    "browserify": "^6.2.0",
-    "envify": "^3.0.0",
-    "jest-cli": "~0.1.17",
-    "reactify": "^0.15.2",
-    "uglify-js": "~2.4.15",
-    "watchify": "^2.1.1",
-    "eslint": "^0.24.1",
-    "eslint-plugin-react": "^3.0.0"
+    "browserify": "^11.0.1",
+    "envify": "^3.4.0",
+    "babelify": "^6.1.3",
+    "uglify-js": "^2.4.24",
+    "watchify": "^3.3.1",
+    "eslint": "^1.1.0",
+    "eslint-plugin-react": "^3.2.3"
   },
   "scripts": {
     "start": "watchify --extension=jsx -o ../static/js/bundle.js -v -d ./**/*.jsx",
@@ -28,7 +27,7 @@
   },
   "browserify": {
     "transform": [
-      "reactify",
+      ["babelify", {"blacklist": ["strict"]}],
       "envify"
     ]
   },
